fix pac-man issues, with restruct for the project

+++ src/objects/entities/Entity.py
@@ -1,9 +1,9 @@
 class Entity(object):
 
-    def __init__(self, step: int, speed: int, position: list):
+    def __init__(self, step: float, speed: int, position: list):
         self.step = step
         self.speed = speed
-        self.position = list(position)  # TODO fix default value should be immutable
+        self.position = list(position)
 
     def x_offset(self, offset: int):
         self.position[1] += offset
